+++ pkgs/native_toolchain_c/test/cbuilder/cbuilder_cross_android_test.dart
@@ -17,6 +17,8 @@ void main() {
     Target.androidArm64,
     Target.androidIA32,
     Target.androidX64,
+    // TODO(rmacnak): Enable when stable NDK 27 is available.
+    // Target.androidRiscv64,
   ];
 
   const readElfMachine = {
@@ -24,6 +26,7 @@ void main() {
     Target.androidArm64: 'AArch64',
     Target.androidIA32: 'Intel 80386',
     Target.androidX64: 'Advanced Micro Devices X86-64',
+    Target.androidRiscv64: 'RISC-V',
   };
 
   const objdumpFileFormat = {
@@ -31,6 +34,7 @@ void main() {
     Target.androidArm64: 'elf64-littleaarch64',
     Target.androidIA32: 'elf32-i386',
     Target.androidX64: 'elf64-x86-64',
+    Target.androidRiscv64: 'elf64-littleriscv',
   };
 
   /// From https://docs.flutter.dev/reference/supported-platforms.